Move AQTTargetGateset to a dedicated module
